+++ src/Testing/MakesGraphQLRequests.php
@@ -151,7 +151,7 @@ trait MakesGraphQLRequests
 
         return Arr::first(
             $results,
-            function (array $result) use ($name): bool {
+            static function (array $result) use ($name): bool {
                 return $result['name'] === $name;
             }
         );