+++ CHANGELOG.md
@@ -13,6 +13,8 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 - Add `@hash` directive which uses Laravel's hashing configuration https://github.com/nuwave/lighthouse/pull/1200
 - Add option `passOrdered` to `@method` to pass just the arguments as ordered parameters https://github.com/nuwave/lighthouse/pull/1208
+- Add support to extend `input`, `interface` and `enum` types https://github.com/nuwave/lighthouse/pull/1203
+- Implement `streamGraphQL()` helper in `\Nuwave\Lighthouse\Testing\MakesGraphQLRequestsLumen` 
 
 ### Deprecated
 
@@ -37,7 +39,6 @@ You can find and compare releases at the [GitHub release page](https://github.co
   get the arguments passed through a client directive https://github.com/nuwave/lighthouse/pull/1184
 - Add `streamGraphQL()` helper method to `\Nuwave\Lighthouse\Testing\MakesGraphQLRequests` for
   simple testing of streamed responses, such as `@defer` https://github.com/nuwave/lighthouse/pull/1184
-- Add support to extend `input`, `interface` and `enum` types https://github.com/nuwave/lighthouse/pull/1203
 
 ### Fixed