+++ CHANGELOG.md
@@ -14,7 +14,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Add `@hash` directive which uses Laravel's hashing configuration https://github.com/nuwave/lighthouse/pull/1200
 - Add option `passOrdered` to `@method` to pass just the arguments as ordered parameters https://github.com/nuwave/lighthouse/pull/1208
 - Add support to extend `input`, `interface` and `enum` types https://github.com/nuwave/lighthouse/pull/1203
-- Implement `streamGraphQL()` helper in `\Nuwave\Lighthouse\Testing\MakesGraphQLRequestsLumen` 
+- Implement `streamGraphQL()` helper in `\Nuwave\Lighthouse\Testing\MakesGraphQLRequestsLumen` https://github.com/nuwave/lighthouse/pull/1222
 
 ### Deprecated